requirements.txt
absl-py==0.9.0
astor==0.8.1
certifi==2019.11.28
click==7.1.1
Flask==1.1.1
Flask-Cors==3.0.8
gast==0.3.3
google-pasta==0.2.0
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.11.1
Keras==2.3.1
Markdown==3.2.1
MarkupSafe==1.1.1
numpy==1.18.2
opencv-python==4.1.2.30
Pillow==7.0.0
protobuf==3.11.3
PyYAML==5.3.1
scipy==1.4.1
six==1.14.0
termcolor==1.1.0
Werkzeug==1.0.0
wrapt==1.12.1

